Initial review feedback.
- Move InputAwareWebView into its own top level class.
- Rename `flutterView` to `containerView` throughout.
- Move the current login in onCreateInputConnection to checkInputConnectionProxy.
- Misc doc cleanup.
